[MTM-25] [#124] fix: vercel 오류 해결 시도
